Backend finalizado

+++ clase1/backend/controllers/tareasControllers.js
@@ -2,7 +2,7 @@ const asyncHandler = require('express-async-handler')
 const Tarea = require('../models/tareasModel')
 
 const getTareas = asyncHandler(async (req, res) => {
-  const tareas = await Tarea.find()
+  const tareas = await Tarea.find({ user: req.user.id })
   res.status(200).json(tareas)
 })
 
@@ -13,7 +13,8 @@ const createTareas = asyncHandler(async (req, res) => {
   }
 
   const tarea = await Tarea.create({
-    descripcion: req.body.descripcion
+    descripcion: req.body.descripcion,
+    user: req.user.id
   })
 
   res.status(201).json(tarea)
@@ -27,8 +28,14 @@ const updateTareas = asyncHandler(async (req, res) => {
     throw new Error('La tarea no existe')
   }
 
-  const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
-  res.status(200).json(tareaUpdated)
+  // Nos aseguramos que la tarea pertenezca al usuario logueado, es decir, el del token
+  if (tarea.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('Usuario no autorizado')
+  } else {
+    const tareaUpdated = await Tarea.findByIdAndUpdate(req.params.id, req.body, { new: true })
+    res.status(200).json(tareaUpdated)
+  }
 })
 
 const deleteTareas = asyncHandler(async (req, res) => {
@@ -39,9 +46,15 @@ const deleteTareas = asyncHandler(async (req, res) => {
     throw new Error('La tarea no existe')
   }
 
-  await Tarea.deleteOne(tarea)
-  // const tareaDeleted = await Tarea.findByIdAndDelete(req.params.id)
-  res.status(200).json({ id: req.params.id })
+  // Nos aseguramos que la tarea pertenezca al usuario logueado, es decir, el del token
+  if (tarea.user.toString() !== req.user.id) {
+    res.status(401)
+    throw new Error('Usuario no autorizado')
+  } else {
+    await Tarea.deleteOne(tarea)
+    // const tareaDeleted = await Tarea.findByIdAndDelete(req.params.id)
+    res.status(200).json({ id: req.params.id })
+  }
 })
 
 module.exports = {
